Organize layout classes
